Resolving missing targets data quality issue
