+++ decode/clarity.ts
@@ -1,11 +1,11 @@
 import version from "../src/core/version";
 import { Event, Payload, Token } from "../types/data";
-import { MetricEvent, PageEvent, PingEvent, SummaryEvent, TagEvent, UploadEvent } from "../types/decode/data";
+import { MetricEvent, PageEvent, PingEvent, SummaryEvent, TagEvent, TargetEvent, UploadEvent } from "../types/decode/data";
 import { DecodedEvent, DecodedPayload } from "../types/decode/decode";
 import { ImageErrorEvent, ScriptErrorEvent } from "../types/decode/diagnostic";
 import { InputChangeEvent, PointerEvent, ResizeEvent, ScrollEvent } from "../types/decode/interaction";
 import { SelectionEvent, UnloadEvent, VisibilityEvent } from "../types/decode/interaction";
-import { BoxModelEvent, DocumentEvent, DomEvent, HashEvent, ResourceEvent, TargetEvent } from "../types/decode/layout";
+import { BoxModelEvent, DocumentEvent, DomEvent, HashEvent, ResourceEvent } from "../types/decode/layout";
 
 import * as data from "./data";
 import * as diagnostic from "./diagnostic";
@@ -44,6 +44,10 @@ export function decode(input: string): DecodedPayload {
             case Event.Tag:
                 if (payload.tag === undefined) { payload.tag = []; }
                 payload.tag.push(data.decode(entry) as TagEvent);
+                break;
+            case Event.Target:
+                if (payload.target === undefined) { payload.target = []; }
+                payload.target.push(data.decode(entry) as TargetEvent);
                 break;
             case Event.Metric:
                 if (payload.metric === undefined) { payload.metric = []; }
@@ -90,10 +94,6 @@ export function decode(input: string): DecodedPayload {
             case Event.Visibility:
                 if (payload.visibility === undefined) { payload.visibility = []; }
                 payload.visibility.push(interaction.decode(entry) as VisibilityEvent);
-                break;
-            case Event.Target:
-                if (payload.target === undefined) { payload.target = []; }
-                payload.target.push(layout.decode(entry) as TargetEvent);
                 break;
             case Event.BoxModel:
                 if (payload.boxmodel === undefined) { payload.boxmodel = []; }
