+++ src/interaction/change.ts
@@ -3,7 +3,6 @@ import { InputChangeData } from "@clarity-types/interaction";
 import { bind } from "@src/core/event";
 import mask from "@src/core/mask";
 import { get } from "@src/layout/dom";
-import * as target from "@src/layout/target";
 import encode from "./encode";
 
 export let data: InputChangeData;
@@ -16,7 +15,6 @@ function recompute(evt: UIEvent): void {
     let input = evt.target as HTMLInputElement;
     let value = get(input);
     if (input && value) {
-        target.observe(value.id);
         data = { target: value.id, value: value.metadata.masked ? mask(input.value) : input.value };
         encode(Event.InputChange);
     }
