+++ src/layout/index.ts
@@ -3,7 +3,6 @@ import * as discover from "@src/layout/discover";
 import * as doc from "@src/layout/document";
 import * as dom from "@src/layout/dom";
 import * as mutation from "@src/layout/mutation";
-import * as target from "@src/layout/target";
 
 export function start(): void {
     doc.reset();
@@ -11,13 +10,11 @@ export function start(): void {
     mutation.start();
     discover.start();
     boxmodel.reset();
-    target.reset();
 }
 
 export function end(): void {
     dom.reset();
     mutation.end();
     boxmodel.reset();
-    target.reset();
     doc.reset();
 }
