+++ types/decode/data.d.ts
@@ -1,4 +1,4 @@
-import { MetricData, PageData, PingData, SummaryData, TagData, UploadData } from "../data";
+import { MetricData, PageData, PingData, SummaryData, TagData, TargetData, UploadData } from "../data";
 import { PartialEvent } from "./core";
 
 /* Data Events */
@@ -7,7 +7,8 @@ export interface PageEvent extends PartialEvent { data: PageData; }
 export interface PingEvent extends PartialEvent { data: PingData; }
 export interface SummaryEvent extends PartialEvent { data: SummaryData[]; }
 export interface TagEvent extends PartialEvent { data: TagData; }
+export interface TargetEvent extends PartialEvent { data: TargetData[]; }
 export interface UploadEvent extends PartialEvent { data: UploadData; }
 export interface DataEvent extends PartialEvent {
-    data: MetricData | PageData | PingData | SummaryData[] | TagData | UploadData;
+    data: MetricData | PageData | PingData | SummaryData[] | TagData | UploadData | TargetData[];
 }
