+++ types/decode/decode.d.ts
@@ -1,9 +1,9 @@
 import { Envelope, Event, MetricData, PageData, PingData, SummaryData, TagData, UploadData } from "../data";
-import { DataEvent, MetricEvent, PageEvent, PingEvent, SummaryEvent, TagEvent, UploadEvent } from "./data";
+import { DataEvent, MetricEvent, PageEvent, PingEvent, SummaryEvent, TagEvent, TargetEvent, UploadEvent } from "./data";
 import { DiagnosticEvent, ImageErrorEvent, ScriptErrorEvent } from "./diagnostic";
 import { InputChangeEvent, InteractionEvent, PointerEvent, ResizeEvent } from "./interaction";
 import { ScrollEvent, SelectionEvent, UnloadEvent, VisibilityEvent } from "./interaction";
-import { BoxModelEvent, DocumentEvent, DomEvent, HashEvent, LayoutEvent, ResourceEvent, TargetEvent } from "./layout";
+import { BoxModelEvent, DocumentEvent, DomEvent, HashEvent, LayoutEvent, ResourceEvent } from "./layout";
 
 export type DecodedEvent = DataEvent | DiagnosticEvent | InteractionEvent | LayoutEvent;
