+++ types/decode/layout.d.ts
@@ -1,4 +1,4 @@
-import { Attributes, BoxModelData, DocumentData, HashData, ResourceData, TargetData } from "../layout";
+import { Attributes, BoxModelData, DocumentData, HashData, ResourceData } from "../layout";
 import { PartialEvent } from "./core";
 
 export interface BoxModelEvent extends PartialEvent { data: BoxModelData[]; }
@@ -6,9 +6,8 @@ export interface HashEvent extends PartialEvent { data: HashData[]; }
 export interface DocumentEvent extends PartialEvent { data: DocumentData; }
 export interface DomEvent extends PartialEvent { data: DomData[]; }
 export interface ResourceEvent extends PartialEvent { data: ResourceData[]; }
-export interface TargetEvent extends PartialEvent { data: TargetData[]; }
 export interface LayoutEvent extends PartialEvent {
-    data: BoxModelData[] | HashData[] | DocumentData | DomData[] | ResourceData[] | TargetData[];
+    data: BoxModelData[] | HashData[] | DocumentData | DomData[] | ResourceData[];
 }
 
 /* Event Data */
